Replace old SMAP test for the current one
